More deduplication, and state file organization
