+++ www/State/SongPlayback.ts
@@ -9,10 +9,29 @@ import {
   typecheck,
 } from '@freik/typechk';
 import { atom } from 'jotai';
+import { atomWithReset } from 'jotai/utils';
 import { IgnoreItem } from 'www/Constants';
 import { Song, SongKey, StorageId } from 'www/Shared/CommonTypes';
+import { MakeSortKey } from 'www/Tools/Sorting';
 import { songByKey } from './Songs';
 import { atomWithMainStorage } from './Storage';
+
+export const isMiniplayerState = atom(false);
+
+// This is the sort for the current playlist
+export const nowPlayingSortState = atomWithReset(MakeSortKey([''], ['lyrnt']));
+
+// The currently selected song(s) to display details for
+export const songDetailState = atomWithReset<Set<SongKey>>(new Set<SongKey>());
+
+// The # of recently added songs to show for a few seconds
+export const recentlyQueuedState = atomWithReset<number>(0);
+
+// Is the 'display message' showing
+export const displayMessageState = atom<boolean>(false);
+
+// Are we displaying the play order?
+export const playOrderDisplayingState = atom<boolean>(false);
 
 export const mutedState = atomWithMainStorage(StorageId.Mute, false, isBoolean);
 export const volumeState = atomWithMainStorage(StorageId.Volume, 0.5, isNumber); // Volume as a percentage (0-100)
