+++ www/Tools/Notifier.tsx
@@ -1,8 +1,10 @@
 import { MessageBar, MessageBarType } from '@fluentui/react';
 import { useAtom } from 'jotai';
 import { ReactElement, useEffect } from 'react';
-
-import { displayMessageState, recentlyQueuedState } from 'www/State/Local';
+import {
+  displayMessageState,
+  recentlyQueuedState,
+} from 'www/State/SongPlayback';
 
 import './styles/Notifier.css';
